Rename test extensions env variable and fs module; split tests according to module
